include python requirements and graph updates
include the requirements.txt file for python so that dependencies can be
installed via pip install -r requirements.txt

create a boolean tag for draftgraph so that if we have no initial api
connection we will not load the graphs. loading empty graphs cause
errors/site crashing when hovering the graph.

Add legend to the graphs.